Added "Export ZIP" function to main page (#47)
Including:

Enabling/disabling the function based on a checkbox in the popup.

Using the tags or description text from the DOM, if any, before using the
saved information from ShaderToy.  This is so readme.txt can be adjusted
(to a certain extent) before download.

When loading scripts, instead of making an AJAX request and stuffing the
text in the body, loading the script directly as a file.  That way the JS
will be listed in the "Sources" panel of the Chrome Developer Tools,
which makes it easier to find and debug.

Added license information for the dependent projects to LICENSE.md.

eslint and jshint checks are clean.